Executable demo to show the basic idea of the engine
VERY BRIEF demo. One box "falling" with a downward velocity collides into a stack of 6 boxes in a zeroG environment, causing them to all scatter until the velocity damping kills all forward momentum. Not a true zeroG environment, but the damping adds to the feeling of getting the forces feeling realistic.